% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_manipulation.R
\name{add_facet}
\alias{add_facet}
\title{Add new facet to model}
\usage{
add_facet(model, facet, property_def = list(), name_column = TRUE)
}
\arguments{
\item{model}{A model}

\item{facet}{Name of the facet to add}

\item{property_def}{List of properties with their type}

\item{name_column}{Whether to add a unique name column}
}
\value{
A modified model
}
\description{
Add new facet to model
}
