% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{add_fragment}
\alias{add_fragment}
\title{Add a fragment to another one}
\usage{
add_fragment(fragment1, fragment2)
}
\arguments{
\item{fragment1}{Fragment to add to}

\item{fragment2}{Fragment to add}
}
\value{
A combination of fragment1 and fragment2
}
\description{
Add a fragment to another one
}
