% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_facet.R
\name{add_fragments}
\alias{add_fragments}
\title{Add a fragment to another one}
\usage{
add_fragments(fragment1, fragment2)
}
\arguments{
\item{fragment1}{Fragment to add to}

\item{fragment2}{Fragment to be added}
}
\value{
A combination of fragment1 and fragment2
}
\description{
Add a fragment to another one
}
