% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{algebraic}
\alias{algebraic}
\title{Create an algebraic relationship}
\usage{
algebraic(definition)
}
\arguments{
\item{definition}{The definition}
}
\value{
A algebraic fragment
}
\description{
Create an algebraic relationship
}
