% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{arg2dec}
\alias{arg2dec}
\title{Interpret function argument as declaration}
\usage{
arg2dec(arg)
}
\arguments{
\item{arg}{Variable to be interpreted}
}
\value{
A declaration or an error message if the variable could not be interpreted as a declaration
}
\description{
Helper function to simplify the common task of checking whether an argument could serve as a declaration, raising an error or
doing the conversion.
}
