% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{arg2fml}
\alias{arg2fml}
\title{Interpret function argument as a formula}
\usage{
arg2fml(arg)
}
\arguments{
\item{arg}{Variable to be interpreted}
}
\value{
A formula or an error message if the variable could not be interpreted as such
}
\description{
Helper function to simplify the common task of checking whether an argument could serve as a formula,
raising an error or doing the conversion.
}
