% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{arg_as_declaration}
\alias{arg_as_declaration}
\title{Interpret function argument as declaration}
\usage{
arg_as_declaration(arg)
}
\arguments{
\item{arg}{Variable to be interpreted}
}
\value{
A declaration or an error message if the variable could not be interpreted as a declaration
}
\description{
Helper function to simplify the common task of checking whether an argument could serve as a declaration, raising an error or
doing the conversion.
}
