% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{as_declaration}
\alias{as_declaration}
\alias{as_declaration.character}
\alias{as_declaration.numeric}
\alias{as_declaration.formula}
\title{Conversion to a declaration}
\usage{
as_declaration(x)

\method{as_declaration}{character}(x)

\method{as_declaration}{numeric}(x)

\method{as_declaration}{formula}(x)
}
\arguments{
\item{x}{Object to convert}
}
\value{
A declaration
}
\description{
Conversion to a declaration
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Returns a declaration with the definition corresponding to the parsed character vector or an error if parsing was not successful.

\item \code{numeric}: Returns a declaration with the definition set to the number provided.

\item \code{formula}: Returns a declaration with the identifier and definition taken from the LHS and RHS of provided formula.
}}

\examples{
d1 <- as_declaration("a+b+c")
d2 <- as_declaration(1)
d3 <- as_declaration(y~b+1)
}
