% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration_list.R
\name{as_declaration_list}
\alias{as_declaration_list}
\title{Make a list of declarations}
\usage{
as_declaration_list(l)
}
\arguments{
\item{l}{list to be converted}
}
\value{
list of declaration objects or error if any element could not be converted
}
\description{
Converts each element of a list to a declaration
}
\examples{
list(~cl, "V") \%>\%  as_declaration_list()
}
