% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R
\name{as_equation}
\alias{as_equation}
\alias{as_equation.character}
\alias{as_equation.numeric}
\alias{as_equation.formula}
\title{Conversion to an equation}
\usage{
as_equation(x)

\method{as_equation}{character}(x)

\method{as_equation}{numeric}(x)

\method{as_equation}{formula}(x)
}
\arguments{
\item{x}{Object to convert}
}
\description{
Conversion to an equation
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Returns an equation with the RHS corresponding to the parsed character vector or an error if parsing was not successful.

\item \code{numeric}: Returns an equation with the RHS set to the number provided.

\item \code{formula}: Returns an equation with LHS and RHS taken from the provided formula.
}}

\examples{
eqn1 <- as_equation("a+b+c")
eqn2 <- as_equation(1)
eqn3 <- as_equation(y~b+1)
}
