% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{as_expr}
\alias{as_expr}
\alias{as_expr.formula}
\alias{as_expr.list}
\alias{as_expr.call}
\title{Convert to an expression}
\usage{
as_expr(x)

\method{as_expr}{formula}(x)

\method{as_expr}{list}(x)

\method{as_expr}{call}(x)
}
\arguments{
\item{x}{Object to convert}
}
\description{
Convert to an expression
}
