% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblerr-package.R
\docType{package}
\name{assemblerr-package}
\alias{assemblerr}
\alias{assemblerr-package}
\title{assemblerr: Assembly of Pharmacometric Models}
\description{
Assembly of pharmacometric models from individual components.
}
\author{
\strong{Maintainer}: Sebastian Ueckert \email{sebastian.ueckert@gmail.com} (0000-0002-3712-0255)

}
\keyword{internal}
