% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-options.R
\name{assemblerr_options}
\alias{assemblerr_options}
\title{Options}
\usage{
assemblerr_options(
  prm.use_mu_referencing = FALSE,
  ode.use_special_advans = TRUE,
  ode.use_general_linear_advans = TRUE,
  ode.general_nonlinear_advan = "advan13",
  ode.general_linear_advan = "advan5",
  ode.preferred_trans_routines = c("trans2", "trans4"),
  issues.missing_variables = c("fix-warn", "fix", "ignore", "fail")
)
}
\arguments{
\item{prm.use_mu_referencing}{Use mu-referencing?}

\item{ode.use_special_advans}{Use analytic solution ADVANs?}

\item{ode.use_general_linear_advans}{Use ADVANs for linear ODEs?}

\item{ode.general_nonlinear_advan}{ADVAN to be used for non-linear ODEs}

\item{ode.general_linear_advan}{ADVAN to be used for linear ODEs}

\item{ode.preferred_trans_routines}{Order of TRANS routines to be tried}

\item{issues.missing_variables}{How to handle missing variables}
}
\value{
A list of options
}
\description{
This function creates a list of options for the use with the render function.
}
\details{
The function helps to create properly formatted list that can serve as input to the \verb{options=} argument
of the \code{render()} function.
}
