% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{c,IssueList-method}
\alias{c,IssueList-method}
\title{Combine issues}
\usage{
\S4method{c}{IssueList}(x, ...)
}
\arguments{
\item{x}{An IssueList}

\item{...}{objects to add to the issue list}
}
\description{
Combine issues
}
\keyword{internal}
