% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{compartment}
\alias{compartment}
\title{Create a new compartment}
\usage{
compartment(name, volume = 1)
}
\arguments{
\item{volume}{}
}
\value{
A compartment item
}
\description{
Create a new compartment
}
