% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{dec_combine}
\alias{dec_combine}
\title{Combine two declarations}
\usage{
dec_combine(d1, d2, op = "+", identifier)
}
\arguments{
\item{d1}{Declaration}

\item{d2}{Declaration}

\item{op}{Operation (function) to combine the declarations with}

\item{identifier}{Identifier for the resulting declaration (taken from d1 if missing)}
}
\value{
A declaration with the specified identifier and the defintion resulting from combining d1 with d2 using op
}
\description{
The function allows the combination of two declarations using an arbitrary operation.
}
\examples{
d1 <- declaration(definition = ka*A["depot"])
d2 <- declaration(definition = ke*A["central"])
d3 <- dec_combine(d1, d2, "-", "dA")
}
