% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{dec_funs_subs}
\alias{dec_funs_subs}
\title{Substitute functions in a declaration}
\usage{
dec_funs_subs(d, substitutions)
}
\arguments{
\item{d}{Declaration}

\item{substitutions}{A list of function names with their corresponding substitution}
}
\value{
Declaration with substituted function names in the definition
}
\description{
Substitute functions in a declaration
}
\examples{
d <- declaration("cl", theta*exp(eta))
assemblerr:::dec_funs_subs(d, c(exp = 'log'))
}
\keyword{internal}
