% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{dec_set_def}
\alias{dec_set_def}
\title{Set identifier or definition of a declaration}
\usage{
dec_set_def(d, definition)
}
\arguments{
\item{d}{A declaration}

\item{definition}{An expression for the new definition}

\item{identifier}{The name of the new identifier}
}
\value{
The modified declaration
}
\description{
Set identifier or definition of a declaration
}
\examples{
d <- declaration("cl", theta*exp(eta))
d2 <- dec_set_def(d, "v")
d3 <- dec_set_id(d, theta)
}
