% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{dec_subs}
\alias{dec_subs}
\title{Substitute symbols in a declaration}
\usage{
dec_subs(d, ...)
}
\arguments{
\item{d}{Declaration}

\item{...}{A list of declarations}
}
\value{
Declaration with substituted symbols in the definition
}
\description{
Substitute symbols in a declaration
}
\examples{
d <- declaration("cl", theta*exp(eta))
d1 <- declaration("theta", theta+a)
assemblerr:::dec_subs(d, d1)
}
