% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{dec_vars}
\alias{dec_vars}
\title{List variables and functions of a declaration}
\usage{
dec_vars(d)
}
\arguments{
\item{d}{A declaration}
}
\value{
A character vector of variables or function used in the definition.
}
\description{
List variables and functions of a declaration
}
\examples{
d <- declaration("cl", theta*exp(eta))
dec_vars(d)
dec_funs(d)
}
