% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{declaration}
\alias{declaration}
\title{Create a declaration}
\usage{
declaration(identifier = NULL, definition)
}
\arguments{
\item{identifier}{The name of the variable to be declared}

\item{definition}{An expression}
}
\value{
A declaration object
}
\description{
Declarations are elementary building blocks of a model that provide the definition for an identifier
}
\examples{
d <- declaration("cl", theta*exp(eta))
}
