% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R
\name{equation}
\alias{equation}
\alias{empty_equation}
\title{Create an equation}
\usage{
equation(expr)

empty_equation()
}
\arguments{
\item{expr}{An expression either of the form lhs~rhs or rhs}
}
\value{
An equation object
}
\description{
Equations are elementary building blocks of a model that describe the mathematical relationship bewtween variables
}
\section{Functions}{
\itemize{
\item \code{empty_equation}: Creates an empty equation (without lhs and rhs)
}}

\examples{
eqn1 <- equation(cl~theta*exp(eta))
eqn2 <- equation(~ka*A["central"])
}
