% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{facet_names_to_labels}
\alias{facet_names_to_labels}
\title{Convert facet class names to labels}
\usage{
facet_names_to_labels(names)
}
\arguments{
\item{names}{Character vector of facet class names}
}
\value{
Character vector with class names translated to labels
}
\description{
Convert facet class names to labels
}
\examples{
assemblerr:::facet_names_to_labels(c("ParameterFacet", "InputVariableFacet"))
}
