% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{fml_has_valid_lhs}
\alias{fml_has_valid_lhs}
\title{Test if LHS of formula is valid}
\usage{
fml_has_valid_lhs(fml, allow_null = TRUE)
}
\arguments{
\item{fml}{A formula}
}
\description{
Test if LHS of formula is valid
}
