% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_facet.R
\name{get_by_name}
\alias{get_by_name}
\title{Retrieve an entry by name}
\usage{
get_by_name(model, facet, name)
}
\arguments{
\item{facet}{Name of the facet}

\item{name}{Name to retrieve}

\item{fragment}{A fragment}
}
\value{
List representation of the entry found or an empty list
}
\description{
Retrieve an entry by name
}
