% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_values.R
\name{get_obs_pv}
\alias{get_obs_pv}
\title{Get parameter value for observational models}
\usage{
get_obs_pv(m, obs, prmz)
}
\arguments{
\item{m}{Model}

\item{obs}{Observation}

\item{prmz}{Requested parameterization}
}
\value{
Parameter values as a named vector
}
\description{
Get parameter value for observational models
}
