% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_values.R
\name{get_pv}
\alias{get_pv}
\title{Get parameter values}
\usage{
get_pv(m, prm, prmz)
}
\arguments{
\item{m}{Model}

\item{prm}{Parameter}

\item{prmz}{Parameterization}
}
\value{
Parameter values as a named vector in requested parameterization or a vector of NA values if
something went wrong.
}
\description{
helpful: (https://sites.google.com/site/probonto/download)
}
