% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_values.R
\name{get_pv_log_normal}
\alias{get_pv_log_normal}
\alias{get_pv_normal}
\title{Get parameter values}
\usage{
get_pv_log_normal(m, prm_name, prmz)

get_pv_normal(m, prm_name, prmz)
}
\arguments{
\item{m}{Model}

\item{prm_name}{Parameter name}

\item{prmz}{Parameterization}
}
\description{
Get parameter values
}
