% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_facet.R
\name{has_facet}
\alias{has_facet}
\title{Test if a model has a facet}
\usage{
has_facet(model, facet)
}
\arguments{
\item{model}{A model}

\item{facet}{Name of the facet}
}
\value{
TRUE/FALSE
}
\description{
Test if a model has a facet
}
\keyword{internal}
