% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{index_subs_dec}
\alias{index_subs_dec}
\title{Substitute the index names in a declaration}
\usage{
index_subs_dec(d, array_name, substitutions)
}
\arguments{
\item{d}{Declaration}

\item{array_name}{Name of the array the index names should be replaced for}

\item{substitutions}{List of subsitutions}
}
\value{
Declaration with replaced index names
}
\description{
Substitute the index names in a declaration
}
\examples{
d <- declaration("dA", ka*A["depot"]-ke*A["central"])
index_subs_dec(d, "A", list(depot = 1, central = 2))
}
