% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{interp}
\alias{interp}
\title{Perform string interpolation with pluralization}
\usage{
interp(..., .envir = parent.frame())
}
\arguments{
\item{...}{Character vectors to interpolate}

\item{.envir}{Environment for lookup}
}
\value{
A character vector
}
\description{
Perform string interpolation with pluralization
}
\examples{
x <- 1:10
assemblerr:::interp("x has elements {x}")
}
\keyword{internal}
