% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{is_valid_lhs}
\alias{is_valid_lhs}
\title{Test if an expression is a valid LHS for a declaration}
\usage{
is_valid_lhs(expr)
}
\arguments{
\item{expr}{an expression}
}
\value{
TRUE/FALSE
}
\description{
Test if an expression is a valid LHS for a declaration
}
\keyword{internal}
