% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_facet.R
\name{item}
\alias{item}
\title{Create an individual entry fragment}
\usage{
item(facet, ...)
}
\arguments{
\item{facet}{The name of the facet this item should be added too}

\item{...}{Properties of the item}
}
\value{
A fragment consisting of just one entry
}
\description{
Create an individual entry fragment
}
\keyword{internal}
