% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_facet.R
\name{list_facets}
\alias{list_facets}
\title{List available facets}
\usage{
list_facets(model)
}
\arguments{
\item{model}{The model}
}
\value{
A character vector
}
\description{
The function lists all facets of a model.
}
