% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{md_doc_links_for_package_functions}
\alias{md_doc_links_for_package_functions}
\title{Create documentation links functions}
\usage{
md_doc_links_for_package_functions(pattern)
}
\arguments{
\item{pattern}{Regular expression to select the package functions}
}
\value{
Character vector of length 1
}
\description{
Creates a string with the Markdown code linking to the doc for all function that match the specifed pattern.
}
\examples{
md_doc_links_for_package_functions("^prm_")
}
\keyword{internal}
