% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{meta_tag}
\alias{meta_tag}
\title{Create a meta tag facet}
\usage{
meta_tag(name, value)
}
\arguments{
\item{name}{Tag name}

\item{value}{Tag value}
}
\value{
A meta tag facet
}
\description{
Create a meta tag facet
}
