% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{General model}
\usage{
model()
}
\value{
A general pharmacometric model
}
\description{
This building block creates the basis for a general pharmacometric model, a flexible but verbose model type.
}
\details{
The object created is just an empty structure to which different building blocks need to be added. In a general model, each building block
can modify one or more of the following facets
\itemize{
\item \link[=parameter-model]{Parameters}
\item \link[=observation-model]{Observations}
\item \link[=algebraic]{Algebraic relationships}
\item \link[=compartment]{Compartments}
\item \link[=flow]{Flows}
\item \link[=input_variables]{Input variables}
\item Meta data
}

The more specialized \code{\link[=pk_model]{pk_model()}} is converted to a general model during the rendering process.
}
\examples{
m <- model() +
    input_variable("dose") +
    prm_log_normal("emax") +
    prm_log_normal("ed50") +
    obs_additive(eff~emax*dose/(ed50+dose))
render(m)
}
