% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\docType{class}
\name{NamedComponent-class}
\alias{NamedComponent-class}
\alias{NamedComponent}
\title{S4 class to represent a named component}
\description{
This is the base class for components with a name
}
\section{Slots}{

\describe{
\item{\code{name}}{A length-one character vector}
}}

