% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-list.R
\docType{class}
\name{NamedComponentList-class}
\alias{NamedComponentList-class}
\alias{NamedComponentList}
\title{S4 class to represent a list of named components}
\description{
This is the base class for a list of named components.
}
