% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{names,NamedFacet-method}
\alias{names,NamedFacet-method}
\title{Access facet names}
\usage{
\S4method{names}{NamedFacet}(x)
}
\arguments{
\item{x}{a named facet}
}
\description{
Access facet names
}
\keyword{internal}
