% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration-creation.R
\name{new_declaration}
\alias{new_declaration}
\title{`declaration` constructor}
\usage{
new_declaration(identifier = list(), definition = list())
}
\arguments{
\item{identifier}{List of expressions}

\item{definition}{List of expressions}
}
\value{
An assemblerr_declaration object
}
\description{
The internal constructor for a declaration vector. The user-facing version is `declaration`.
}
\details{
The arguments `identifier` and `definition` are lists of R expressions. For `identifier` only symbols, array expressions
(e.g., theta[1]), or NULL are permitted.
}
\keyword{internal}
