% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_facet.R
\name{new_facet}
\alias{new_facet}
\title{Model facets}
\usage{
new_facet(facet_name = character(), x = list(),
  .id_columns = character())
}
\arguments{
\item{name}{Name of the facet to add}

\item{property_definitions}{List of properties with their type}

\item{default_definitions}{List of properties that will be added to the facet by default with their type}

\item{id_columns}{Column names that together yield a unique identifier of the facet entries}
}
\value{
A facet
}
\description{
Model facets
}
\examples{
f <- new_facet("theta", list(initial = numeric(), lbound = numeric(), ubound = numeric()))
}
\keyword{internal}
