% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model.R
\name{nm_pk}
\alias{nm_pk}
\alias{nm_des}
\alias{nm_error}
\title{Create model code facet}
\usage{
nm_pk(name, statement)

nm_des(name, statement)

nm_error(name, statement)
}
\arguments{
\item{name}{Facet name}

\item{statement}{Code statement}
}
\value{
A facet
}
\description{
Create model code facet
}
