% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model.R
\name{nm_theta}
\alias{nm_theta}
\alias{nm_omega}
\alias{nm_sigma}
\title{Create facet for initial values}
\usage{
nm_theta(name, initial = NA, lbound = -Inf, ubound = Inf)

nm_omega(name, initial = NA)

nm_sigma(name, initial = NA)
}
\arguments{
\item{name}{Parameter name}

\item{initial}{Initial value}

\item{lbound}{Lower bound}

\item{ubound}{Upper bound}
}
\value{
Facet
}
\description{
Create facet for initial values
}
