% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\name{obs_combined}
\alias{obs_combined}
\alias{obs_additive}
\alias{obs_proportional}
\title{Observation}
\usage{
obs_combined(prediction, name)

obs_additive(prediction, name)

obs_proportional(prediction, name)
}
\arguments{
\item{prediction}{A declaration defining the observed variable}

\item{name}{A name for the observation (automatically derived if missing)}
}
\description{
These building blocks specify an observational model describing the observed variable as well as how an observation is expected to diverge from
the model (i.e, the residual unexplained variability model - RUV model).
}
\section{Functions}{
\itemize{
\item \code{obs_combined}: Observation following a combined error model \eqn{y = f + f \epsilon_1 + \epsilon_2}

\item \code{obs_additive}: Observation following an additive error model \eqn{y = f + \epsilon_1}

\item \code{obs_proportional}: Observation following a proportional error model \eqn{y = f + f \epsilon_1}
}}

\examples{
# additve RUV model for observing the variable WT
m <- model() +
  prm_log_normal("wt") +
  obs_additive(~wt)

# EMAX dose-response model with proportional RUV
m2 <- model() +
  input_variable("dose") +
  prm_no_var("emax") +
  prm_no_var("ed50") +
  obs_proportional(effect~emax*dose/(ed50+dose))
}
