% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\name{obs_additive}
\alias{obs_additive}
\alias{obs_proportional}
\alias{obs_combined}
\alias{observation-model}
\title{Observation model}
\usage{
obs_additive(prediction, name)

obs_proportional(prediction, name)

obs_combined(prediction, name)
}
\arguments{
\item{prediction}{A declaration defining the observed variable}

\item{name}{A name for the observation (automatically derived if missing)}
}
\description{
These building blocks specify an observational model describing the observed variable as well as how an observation is expected to diverge from
the model (i.e, the residual unexplained variability model).
}
\details{
The RUV model is choosen by selecting one of the predefined functions.
\itemize{
  \item \code{obs_additive}: Observation following an additive error model \eqn{y = f + \epsilon_1}
  \item \code{obs_proportional}: Observation following a proportional error model \eqn{y = f + f \epsilon_1}
  \item \code{obs_combined}: Observation following a combined error model \eqn{y = f + f \epsilon_1 + \epsilon_2}
}
The first argument of these functions requires the definition of the actual prediction from the model. It can be specified in a number of different ways
\itemize{
  \item the name of a variable in the model: \code{obs_additive("effect")}
  \item a compartment concentration: \code{obs_additive(~C["central"])}
  \item an equation: \code{obs_additive(~base+slp*time)}
}
If the definition contains a variable name on the left-hand side (as in \code{conc~C["central"]}), the variable will appear in the generated model code.
The latter is useful to make the model code more readable if the prediction is defined as a long equation.

The second function argument "name" will be used to support multiple observations but is currently not useful.
}
\examples{
# additve RUV model for observing the variable WT
m <- model() +
  prm_log_normal("wt") +
  obs_additive(~wt)

# EMAX dose-response model with proportional RUV
m2 <- model() +
  input_variable("dose") +
  prm_no_var("emax") +
  prm_no_var("ed50") +
  obs_proportional(effect~emax*dose/(ed50+dose))
}
