% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{parameter}
\alias{parameter}
\title{Create a new model parameter}
\usage{
parameter(name, type = "log-normal")
}
\arguments{
\item{type}{}
}
\value{
A parameter fragment
}
\description{
Create a new model parameter
}
