% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{parameter_values}
\alias{parameter_values}
\title{Set values for parameters}
\usage{
parameter_values(...)
}
\arguments{
\item{...}{parameter names and values}
}
\description{
Set values for parameters
}
\examples{
vals <- parameter_values(cl = c(tv = 5, var = 0.3))
}
