% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_component.R
\name{pk_absorption_zo}
\alias{pk_absorption_zo}
\title{Zero-order absorption model}
\usage{
pk_absorption_zo(prm_mat = prm_log_normal("mat", median = 0.5, var_log = 0.1))
}
\arguments{
\item{prm_mat}{Parameter building block}
}
\value{
An absorption building block
}
\description{
Zero-order absorption model
}
\seealso{
Other PK absorption models: 
\code{\link{pk_absorption_zo_lag}()}
}
\concept{PK absorption models}
