% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_component.R
\name{pk_absorption_zo_lag}
\alias{pk_absorption_zo_lag}
\title{Zero-order absorption with lag-time model}
\usage{
pk_absorption_zo_lag(
  prm_mat = prm_log_normal("mat", median = 0.5, var_log = 0.1),
  prm_mdt = prm_log_normal("mdt", median = 0.5, var_log = 0.1)
)
}
\arguments{
\item{prm_mat}{Parameter building block}
}
\value{
An absorption building block
}
\description{
Zero-order absorption with lag-time model
}
\seealso{
Other PK absorption models: 
\code{\link{pk_absorption_zo}()}
}
\concept{PK absorption models}
