% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_component.R
\name{pk_distribution_2cmp}
\alias{pk_distribution_2cmp}
\title{PK distribution 2 compartment}
\usage{
pk_distribution_2cmp(
  prm_vc = prm_log_normal("vc", median = 100, var_log = 0.1),
  prm_vp = prm_log_normal("vp", median = 5, var_log = 0.1),
  prm_q = prm_log_normal("q", median = 50, var_log = 0.1)
)
}
\arguments{
\item{prm_vc}{Parameter model for the central volume of distribution}

\item{prm_vp}{Parameter model for the peripheral volume of distribution}

\item{prm_q}{Parameter model for the inter-compartmental clearance}
}
\description{
This building block declares a two compartment distribution component for a pharmacokinetic model.
}
\details{
\subsection{PK components}{
\subsection{Parameter models}{
}

}
}
\seealso{
\code{\link[=pk_model]{pk_model()}} for the creation of PK models

Other distribution components: 
\code{\link{pk_distribution_1cmp}()}
}
\concept{distribution components}
