% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{+,BuildingBlock,BuildingBlock-method}
\alias{+,BuildingBlock,BuildingBlock-method}
\title{Add building blocks}
\usage{
\S4method{+}{BuildingBlock,BuildingBlock}(e1, e2)
}
\arguments{
\item{e1}{A building block}

\item{e2}{A building block}
}
\description{
Add building blocks
}
\keyword{internal}
