% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{+,BuildingBlock,NULL-method}
\alias{+,BuildingBlock,NULL-method}
\title{Add building blocks}
\usage{
\S4method{+}{BuildingBlock,`NULL`}(e1, e2)
}
\arguments{
\item{e1}{A building block}

\item{e2}{NULL}
}
\description{
Add building blocks
}
\keyword{internal}
