% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{+,Component,Component-method}
\alias{+,Component,Component-method}
\title{Add components}
\usage{
\S4method{+}{Component,Component}(e1, e2)
}
\arguments{
\item{e1}{A component}

\item{e2}{A component}
}
\description{
Add components
}
\keyword{internal}
