% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{render}
\alias{render}
\title{Generate model code}
\usage{
render(
  model,
  filename = NULL,
  target_tool = "nonmem",
  options = assemblerr_options()
)
}
\arguments{
\item{model}{A model object}

\item{filename}{Name of the model file to create or NULL}

\item{target_tool}{Name of the target tool (currently only 'nonmem')}

\item{options}{List of options for model generation}
}
\description{
This function generates NONMEM code for a model object. The generated code will be written to the file specified by 'filename' or printed to the
console if the argument is set to NULL.
}
