% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{set_definition}
\alias{set_definition}
\alias{set_identifier}
\title{Set identifier or definition of a declaration}
\usage{
set_definition(d, definition)

set_identifier(d, identifier = NULL)
}
\arguments{
\item{d}{A declaration}

\item{definition}{An expression for the new definition}

\item{identifier}{The name of the new identifier}
}
\value{
The modified declaration
}
\description{
Set identifier or definition of a declaration
}
\examples{
d <- declaration("cl", theta*exp(eta))
d2 <- set_identifier(d, "v")
d3 <- set_definition(d, theta)
}
