% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{statement}
\alias{statement}
\title{One or several imperative instructions}
\usage{
statement(...)
}
\arguments{
\item{...}{}
}
\value{
A statement
}
\description{
One or several imperative instructions
}
