% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{subs_dec}
\alias{subs_dec}
\title{Substitute symbols in a declaration}
\usage{
subs_dec(d, ...)
}
\arguments{
\item{d}{Declaration}

\item{...}{A list of symbols and their replacement}
}
\value{
Declaration with substituted symbols in the definition
}
\description{
Substitute symbols in a declaration
}
\examples{
d <- declaration("cl", theta*exp(eta))
subs_dec(d, eta = eta+a)
}
