% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{transform_ast}
\alias{transform_ast}
\title{Modify AST}
\usage{
transform_ast(node, transformer, ...)
}
\arguments{
\item{node}{A language node}

\item{transformer}{A transformer function}

\item{...}{Additional arguments to the transformer function}
}
\value{
The transformed language node
}
\description{
This recursive function is the work-horse for all expression transformations. It takes a language node and a transformer function,
and applies the transformer recursivly to the node and all its child nodes.
}
