% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{tsk_estimation}
\alias{tsk_estimation}
\alias{tsk_output}
\alias{tsk_output_xpose4}
\title{Model tasks}
\usage{
tsk_estimation(
  algorithm = c("foce", "foce-inter", "foce-no-inter", "fo", "imp", "saem"),
  se = FALSE,
  target_options = list()
)

tsk_output(filename = "sdtab", variables)

tsk_output_xpose4()
}
\arguments{
\item{algorithm}{The estimation algorithm to use for the task}

\item{se}{Whether to calculate parameter uncertainties}

\item{target_options}{List of additional options that should be passed to
NONMEM}

\item{filename}{The filename for the output file}

\item{variables}{The model variables that should be output}
}
\description{
These functions define model tasks that can be combined using \code{+} and then be included during rendering.
\itemize{
\item \code{tsk_estimation()} adds a parameter estimation task allowing to specify
the estimation algorithm to be used and whether a $COV step should be run.
\item \code{tsk_ouput()} adds an output task allowing to display model variables
in a table.
\item \code{tsk_ouput_xpose()} adds several tables as output following the xpose4 output
standard.
}
}
\examples{

m <- model() +
  input_variable("dose") +
  prm_log_normal("emax", median = 10, var_log = 0.09) +
  prm_log_normal("ed50", median = 50, var_log = 0.09) +
  algebraic(effect~emax*dose/(ed50 + dose)) +
  obs_proportional(~effect, var_prop = 1)

# add estimation task using importance sampling, covariance step
# and user-defined ISAMPLE option
render(
  model = m,
  tasks = tsk_estimation(
    algorithm = "imp",
    se = TRUE,
    target_options = list(isample=1000)
  )
)

# output model parameters to file 'prms'
render(m, tasks = tsk_output("prms", variables = vars_prms()))
}
