% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{variables}
\alias{variables}
\alias{functions}
\title{List variables and functions of a declaration}
\usage{
variables(d)

functions(d)
}
\arguments{
\item{d}{A declaration}
}
\value{
A character vector of variables or function used in the definition.
}
\description{
List variables and functions of a declaration
}
\examples{
d <- declaration("cl", theta*exp(eta))
variables(d)
functions(d)
}
