% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model.R
\name{write_model}
\alias{write_model}
\title{Write model code to file}
\usage{
write_model(model, filename, directory = getwd())
}
\arguments{
\item{model}{The model}

\item{filename}{The filename for the model file}

\item{directory}{The directory where to save the file}
}
\description{
Write model code to file
}
