% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{xtfrm,IssueList-method}
\alias{xtfrm,IssueList-method}
\title{Auxiliary Function for Sorting and Ranking}
\usage{
\S4method{xtfrm}{IssueList}(x)
}
\arguments{
\item{x}{an R Object}
}
\description{
Auxiliary Function for Sorting and Ranking
}
\keyword{internal}
