% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{algebraic}
\alias{algebraic}
\title{#' Model parameter
#'
#' Defines name and type of a model parameter
#'
#' @param name Name of the paramter
#' @param type Model type to be used for the parameter
#'
#' @return A \code{\link{fragment}} representing a model parameter
#' @export
#' @examples
#' p <- parameter("cl", "log-normal")
parameter <- function(name, type){
  if(name!=make.names(name)) stop("'name' needs to be a valid variable name.")
  if(missing(type)){
    message("No type for the parameter '", name,"' was specified, using 'log-normal' as default.")
    type <- "log-normal"
  }
  item("parameters", name = name, type = type)
}}
\usage{
algebraic(definition)
}
\description{
#' Model parameter
#'
#' Defines name and type of a model parameter
#'
#' @param name Name of the paramter
#' @param type Model type to be used for the parameter
#'
#' @return A \code{\link{fragment}} representing a model parameter
#' @export
#' @examples
#' p <- parameter("cl", "log-normal")
parameter <- function(name, type){
  if(name!=make.names(name)) stop("'name' needs to be a valid variable name.")
  if(missing(type)){
    message("No type for the parameter '", name,"' was specified, using 'log-normal' as default.")
    type <- "log-normal"
  }
  item("parameters", name = name, type = type)
}
}
