% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblerr-package.R
\docType{package}
\name{assemblerr-package}
\alias{assemblerr}
\alias{assemblerr-package}
\title{assemblerr: Assembly of Pharmacometric Models}
\description{
assemblerr is an R package to construct pharmacometric models by
    combining pre-defined model components. It’s intended to simplify the
    specification of complex pharmacometric models and provide a mechanism
    to generate models in an automatic way.
}
\author{
\strong{Maintainer}: Sebastian Ueckert \email{sebastian.ueckert@gmail.com} (\href{https://orcid.org/0000-0002-3712-0255}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Mats O. Karlsson [scientific advisor]
  \item Andrew C. Hooker [scientific advisor]
  \item Rikard Nordgren [scientific advisor]
  \item Simon Carter [reviewer]
  \item Simon Buatois [reviewer]
  \item João A. Abrantes [reviewer]
  \item F. Hoffmann-La Roche Ltd. [funder]
}

}
\keyword{internal}
