% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblerr-package.R
\docType{package}
\name{assemblerr-package}
\alias{assemblerr}
\alias{assemblerr-package}
\title{assemblerr: Assembly of Pharmacometric Models}
\description{
Construct pharmacometric nonlinear mixed effect models by combining predefined model components and automatically generate model code for NONMEM. Models are created by combining parameter and observation models, algebraic relationships, compartments, and flows. Pharmacokinetic models can be assembled from the higher-order components: absorption, distribution, and elimination. The generated code is optimized for performance by recognizing, for example, linear differential equations or differential equations with an analytic solution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UUPharmacometrics/assemblerr}
  \item Report bugs at \url{https://github.com/UUPharmacometrics/assemblerr/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Ueckert \email{sebastian.ueckert@gmail.com} (\href{https://orcid.org/0000-0002-3712-0255}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Mats O. Karlsson [scientific advisor]
  \item Andrew C. Hooker [scientific advisor]
  \item Rikard Nordgren [scientific advisor]
  \item Simon Carter [reviewer]
  \item Simon Buatois [reviewer]
  \item João A. Abrantes [reviewer]
  \item F. Hoffmann-La Roche Ltd. [funder]
}

}
