% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.R
\name{compartment}
\alias{compartment}
\alias{cmp}
\title{Compartment}
\usage{
compartment(name, volume = 1)

cmp(name, volume = 1)
}
\arguments{
\item{name}{Name of the compartment}

\item{volume}{Defintion of the compartment volume as a number, formula or declaration}
}
\value{
A compartment building block
}
\description{
Defines name and volume of a compartment
}
\section{Functions}{
\itemize{
\item \code{cmp}: Is an simple alias for compartment.
}}

\examples{
# add a compartment with name "central" and volume Vc
m <- model() +
 compartment("central", volume = ~vc) +
 flow(~cl*C, from = "central") +
 prm_log_normal("cl") +
 prm_log_normal("vc") +
 obs_additive(conc~C["central"])
}
\seealso{
\code{\link{model}}
}
