% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{compartment}
\alias{compartment}
\alias{cmp}
\title{#' @export
print.model <- function(x,...){
  name <- NA
  title <- get_first(x, "meta_tags", name == "title")$value
  if(!is.null(title)){
    cat('assemblerr model "', title,'":\n')
  } else{
    cat('assemblerr model:\n')
  }
  prms <- x$parameters$name %>% paste(collapse = ", ")
  cat('  - Parameters: ', prms, '\n')
  obs_txt <- x$observations %>%
    dplyr::mutate(txt = paste0("    + \"", .data$name, "\" (", .data$type, ")")) %>%
    purrr::pluck("txt") %>%
    paste(collapse = "\n")
  cat('  - Observations: \n')
  cat(obs_txt, "\n")
  comp_txt <- x$compartments$name %>% paste(collapse = ", ")
  if(nrow(x$compartments)>0) cat('  - Compartments: ', comp_txt, "\n")
  alg_count <- x$algebraics %>% nrow()
  if(alg_count>0) cat('  - Algebraic relationships: ', alg_count, "\n")
}
Compartment}
\usage{
compartment(name, volume = declaration(~1))

cmp(name, volume = declaration(~1))
}
\arguments{
\item{name}{Name of the compartment}

\item{volume}{Defintion of the compartment volume as a number, formula or declaration}
}
\value{
A compartment fragment
}
\description{
Defines name and volume of compartment
}
\section{Functions}{
\itemize{
\item \code{cmp}: Is an simple alias for compartment.
}}

\examples{
# compartment with name "central" and volume Vc
comp1 <- compartment("central", volume = declaration(~Vc))
}
\seealso{
\code{\link{model}}
}
