% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{compartment}
\alias{compartment}
\alias{cmp}
\title{Compartment}
\usage{
compartment(name, volume = 1)

cmp(name, volume = 1)
}
\arguments{
\item{name}{Name of the compartment}

\item{volume}{Defintion of the compartment volume as a number, formula or declaration}
}
\value{
A compartment fragment
}
\description{
Defines name and volume of compartment
}
\section{Functions}{
\itemize{
\item \code{cmp}: Is an simple alias for compartment.
}}

\examples{
# compartment with name "central" and volume Vc
comp1 <- compartment("central", volume = ~Vc)
}
\seealso{
\code{\link{model}}
}
