% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{dcl_id}
\alias{dcl_id}
\alias{dcl_def}
\title{Get declaration fields}
\usage{
dcl_id(dcl)

dcl_def(dcl)
}
\arguments{
\item{dcl}{A declaration}
}
\value{
An expression
}
\description{
These function get identifier or definition fields of a declaration.
}
\keyword{internal}
