% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration.R
\name{dcl_id}
\alias{dcl_id}
\alias{dcl_def}
\title{Access declaration fields}
\usage{
dcl_id(dcl)

dcl_def(dcl)
}
\arguments{
\item{dcl}{An assemblerr_declaration}
}
\value{
An expression
}
\description{
Access declaration fields
}
\keyword{internal}
