% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declaration-creation.R
\name{declaration}
\alias{declaration}
\title{Declaration}
\usage{
declaration(...)
}
\arguments{
\item{...}{List of R formulae with a single symbol on the left-hand side and a valid R expression on the right}
}
\value{
A declaration vector
}
\description{
A declaration is the mathematical definition of a set of variables. It is the lowest level building block for a model
in `assemblerr`. A declaration consists of the variable names being declared (the identifiers) and their definition. The
`declaration` function allows the specification of a declaration using `R` formulae.
}
\examples{
d <- declaration(cl~theta[1]+eta[1])
d2 <- declaration(v=theta[2]*exp(eta[2]))
}
