% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.R
\name{flow}
\alias{flow}
\title{Flows between compartments}
\usage{
flow(definition, from = NA_character_, to = NA_character_)
}
\arguments{
\item{definition}{Declaration of the flow using the variable A (amount in 'from' compartment) and C (concentration in 'from' compartment)}

\item{from}{Name of the source compartment or NA for an inflow without source}

\item{to}{Name of the sink compartment or NA for an ouflow without sink}
}
\value{
A flow building block
}
\description{
Creates a building block describing a flow between compartments.
}
\examples{
f <- model() +
     compartment("depot") +
     compartment("central", volume = ~V)
     flow(declaration(~ka*A), from = "depot", to = "central")
}
