% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{flow}
\alias{flow}
\title{Create a new flow from or to compartments}
\usage{
flow(from = NULL, to = NULL, equation)
}
\arguments{
\item{from}{Name of the source compartment or Null}

\item{to}{Name of the sink compartment or Null}

\item{equation}{An equation describing the flow}
}
\value{
A flow item
}
\description{
Create a new flow from or to compartments
}
