% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartment.R
\name{flow}
\alias{flow}
\title{Flows between compartments}
\usage{
flow(from = NA_character_, to = NA_character_, definition)
}
\arguments{
\item{from}{Name of the source compartment or NULL}

\item{to}{Name of the sink compartment or NULL}

\item{definition}{Declaration of the flow using the special variable A (amount in 'from' compartment) and C (concentration in 'from' compartment)}
}
\value{
A flow fragment
}
\description{
Creates a fragment describing a flow from between compartmens.
}
\examples{
f <- flow(from = "depot", to = "central", definition = declaration(~ka*A))
}
