% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{General model}
\usage{
model()
}
\value{
A general pharmacometric model
}
\description{
\code{model()} creates the foundation for a general pharmacometric model
}
\details{
This function creates a model object, the foundation for a general, software-agnostic description of a pharmacometric model.
The object created is an empty structure. In general, one will want to add components to the model,
then convert it to a software-specific model object and finally create the model code. The following components can be added
to a general model:
}
\examples{
m <- model()+
    obs_additive(eff~emax*dose/(ed50+dose)) +
    prm_log_normal("emax") +
    prm_log_normal("ed50")
}
