% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model.R
\name{nm_model}
\alias{nm_model}
\title{NONMEM model}
\usage{
nm_model(options = options_nm())
}
\value{
An nm_model
}
\description{
\code{nm_model()} creates the foundation for a NONMEM model
}
\details{
This function creates a NONMEM model object, a software-specific version of the general
\code{\link{model}}. Like for the general model,this function only creates the empty
base object which then needs to be filled with components before it can be rendered. The
following components can be added
to a NONMEM model:
\itemize{
   \item \code{\link{nm_problem}}
   \item \code{\link{nm_input}}
   \item \code{\link{nm_pk}}
   \item \code{\link{nm_des}}
   \item \code{\link{nm_error}}
   \item \code{\link{nm_theta}}
   \item \code{\link{nm_omega}}
   \item \code{\link{nm_sigma}}
}
}
