% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model.R
\name{nm_theta}
\alias{nm_theta}
\title{Create facet for initial values}
\usage{
nm_theta(name, initial = 1, lbound = -Inf, ubound = Inf)
}
\arguments{
\item{name}{Parameter name}

\item{initial}{Initial value}

\item{lbound}{Lower bound}

\item{ubound}{Upper bound}
}
\value{
A NONMEM Theta parameter
}
\description{
Create facet for initial values
}
\keyword{internal}
