% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\name{observation}
\alias{observation}
\alias{obs_additive}
\alias{obs_proportional}
\alias{obs_power}
\alias{obs_binary}
\title{Observation model}
\usage{
observation(name, type, options = NULL)

obs_additive(prediction, name)

obs_proportional(prediction, name)

obs_power(prediction, name)

obs_binary(p1, name)
}
\arguments{
\item{name}{The name used to identify the measurement}

\item{type}{The model type used for the observation model}

\item{options}{A \code{\link{declaration}} describing the measurement}

\item{prediction}{Declaration for prediction}

\item{p1}{Declaration for P(Y=1)}
}
\value{
A \code{fragment} representing an observation model
}
\description{
Defines how variables from a model relate to values in the data
}
\examples{
# create an additive error observation model for the concentration from the "central" compartment
c_obs <- observation(name = "conc", type = "additive", options = list(prediction = ~C["central"]))
}
