% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{observation}
\alias{observation}
\title{Create a new observation model}
\usage{
observation(name = "", equation, type)
}
\arguments{
\item{name}{The name used to identify the measurement}

\item{equation}{An equation describing the measurement}

\item{type}{The model type used for the observation model}
}
\value{
A observation model fragment
}
\description{
Observation models relate variables from a model to measured values.
}
\examples{
# create an additive error observation model "conc" for the concentration from the "central" compartment
c_obs <- observation("conc", ~C["central"], "additive")

# create a combined error observation model "eff" for the variable effect
e_obs <- observation("eff", ~effect, "combined")
}
