% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{observation}
\alias{observation}
\title{Create a new observation}
\usage{
observation(name, equation, type)
}
\arguments{
\item{type}{}
}
\value{
A parameter fragment
}
\description{
Create a new observation
}
