% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{parameter}
\alias{parameter}
\title{Model parameter}
\usage{
parameter(name, type, options = NULL)
}
\arguments{
\item{name}{Name of the paramter}

\item{type}{Model type to be used for the parameter}

\item{options}{Options}
}
\value{
A \code{\link{fragment}} representing a model parameter
}
\description{
Defines name and type of a model parameter
}
\examples{
p <- parameter("cl", "log-normal")
}
