% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_model.R
\name{pk_component}
\alias{pk_component}
\alias{pk_distribution_1cmp}
\alias{pk_distribution_2cmp}
\alias{pk_distribution_3cmp}
\alias{pk_elemination_linear}
\alias{pk_elemination_mm}
\alias{pk_absorption_rate_fo}
\alias{pk_absorption_rate_zo}
\alias{pk_absorption_delay_none}
\title{Create PK model components}
\usage{
pk_component(name, class, prm_names)

pk_distribution_1cmp(prm_vc = prm_log_normal("vc"))

pk_distribution_2cmp(prm_vc = prm_log_normal("vc"),
  prm_vp = prm_log_normal("vp"), prm_q = prm_log_normal("q"))

pk_distribution_3cmp()

pk_elemination_linear(prm_cl = prm_log_normal("cl"))

pk_elemination_mm()

pk_absorption_rate_fo()

pk_absorption_rate_zo()

pk_absorption_delay_none()
}
\arguments{
\item{name}{Component name}

\item{type}{Component type}
}
\value{
A pk_model component
}
\description{
Create PK model components
}
