% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rendering.R
\name{render}
\alias{render}
\alias{render_opts_nm}
\title{Render an object}
\usage{
render(object, opts)

render_opts_nm()
}
\arguments{
\item{object}{The object to render}

\item{opts}{Rendering options}
}
\description{
Render an object
}
