% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{statement}
\alias{statement}
\alias{stm}
\title{One or several imperative instructions}
\usage{
statement(...)

stm(...)
}
\arguments{
\item{...}{The statement}
}
\value{
A statement
}
\description{
One or several imperative instructions
}
\section{Functions}{
\itemize{
\item \code{stm}: Alias for statement
}}

